Add tests for the IssuePrintFormatter
This adds tests to cover the methods inside the IssuePrintFormatter.

Note: there is some over-testing in a few of these cases because of how
both Issue and Label are defined. We would need to refactor them to use
interfaces for mocking out this functionality if these tests are to be
pure unit tests. As of now, that seems like a bit too much work for the
scope of this commit.